backend/requirements.txt
fastapi==0.115.14          # 26 Jun 2025 :contentReference[oaicite:0]{index=0}
uvicorn[standard]==0.35.0  # 15 Jun 2025 :contentReference[oaicite:1]{index=1}
SQLAlchemy==2.0.41         # 24 Jun 2025 :contentReference[oaicite:2]{index=2}
psycopg2-binary==2.9.10    # 16 Oct 2024 :contentReference[oaicite:3]{index=3}
alembic==1.16.2            # 16 Jun 2025 :contentReference[oaicite:4]{index=4}
pydantic==2.11.7           # 14 Jun 2025 :contentReference[oaicite:5]{index=5}
python-multipart==0.0.20   # 16 Dec 2024 :contentReference[oaicite:6]{index=6}
python-dotenv==1.1.1       # 24 Jun 2025 :contentReference[oaicite:7]{index=7}
